+++ CHANGELOG.md
@@ -6,6 +6,11 @@ file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
 to structure this file.
 
+## 1.1.0 2019-12-07
+
+- Feat: add title case. Same as start case, except small words like "a" and
+  "the" will remain lower case.
+
 ## 1.0.0
 
 - Initial release